requirements.txt
numpy
scipy
matplotlib
tqdm
scikit-learn
scikit-image
nibabel
SimpleITK
tensorflow-gpu
keras
